Restructure legacy route code
